requirement.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
albumentations==0.4.3
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
dnspython==2.6.1
einops==0.3.0
email_validator==2.2.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.32.0
google-auth-oauthlib==1.0.0
gradio==4.32.2
gradio_client==0.17.0
grpcio==1.64.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgaug==0.2.6
importlib_metadata==8.0.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kornia==0.5.10
lazy_loader==0.4
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
omegaconf==2.1.1
opencv-python==4.1.2.30
opencv-python-headless==4.10.0.84
orjson==3.10.6
packaging==24.1
pandas==2.0.3
pillow==10.4.0
pkgutil_resolve_name==1.3.10
protobuf==5.27.2
pudb==2019.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pyDeprecate==0.3.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==1.4.2
pytorch-msssim==1.0.0
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
ruff==0.5.1
sacremoses==0.1.1
scikit-image==0.20.0
scipy==1.9.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.37.2
streamlit==1.36.0
sympy==1.13.0
tbb==2021.13.0
tenacity==8.5.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
test_tube==0.7.5
tifffile==2023.7.10
tokenizers==0.10.3
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
torch==torch-1.8.1
torch-fidelity==0.3.0
torchmetrics==0.6.0
torchvision==0.9.1
tornado==6.4.1
tqdm==4.66.4
transformers==4.6.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
urwid==2.6.15
uvicorn==0.30.1
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.3
yarl==1.9.4
zipp==3.19.2
